create draft folder
